requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
distro==1.9.0
Django==4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.37.1
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pytz==2024.1
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
